+++ tests/gcontacts/golden/contacts_tool_schemas.json
@@ -275,6 +275,18 @@
         "default": null,
         "description": "Street address (for create/update)."
       },
+      "birthday": {
+        "anyOf": [
+          {
+            "type": "string"
+          },
+          {
+            "type": "null"
+          }
+        ],
+        "default": null,
+        "description": "Birthday as 'YYYY-MM-DD', 'MM-DD' (no year), or 'clear'/'' to remove."
+      },
       "contact_id": {
         "anyOf": [
           {
@@ -574,6 +586,17 @@
             ],
             "default": null
           },
+          "birthday": {
+            "anyOf": [
+              {
+                "type": "string"
+              },
+              {
+                "type": "null"
+              }
+            ],
+            "default": null
+          },
           "email": {
             "anyOf": [
               {
@@ -757,6 +780,17 @@
         "description": "Typed batch-update input for a contact.",
         "properties": {
           "address": {
+            "anyOf": [
+              {
+                "type": "string"
+              },
+              {
+                "type": "null"
+              }
+            ],
+            "default": null
+          },
+          "birthday": {
             "anyOf": [
               {
                 "type": "string"
@@ -1257,7 +1291,8 @@
               "userDefined",
               "relations",
               "biographies",
-              "addresses"
+              "addresses",
+              "birthdays"
             ],
             "type": "string"
           },
@@ -1266,7 +1301,7 @@
           }
         ],
         "default": null,
-        "description": "For \"update\" action \u2014 the single People API field to update across\nall contacts in this batch. Required. Must be one of: names, phoneNumbers,\nemailAddresses, organizations, nicknames, urls, userDefined, relations,\nbiographies, addresses. Using a single field per batch call prevents\nunintentional data loss from a union updateMask overwriting unrelated fields."
+        "description": "For \"update\" action \u2014 the single People API field to update across\nall contacts in this batch. Required. Must be one of: names, phoneNumbers,\nemailAddresses, organizations, nicknames, urls, userDefined, relations,\nbiographies, addresses, birthdays. Using a single field per batch call prevents\nunintentional data loss from a union updateMask overwriting unrelated fields."
       },
       "updates": {
         "anyOf": [
